requirements.txt
attrs==23.2.0
bcrypt==4.1.3
black==24.4.2
blinker==1.8.2
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
distlib==0.3.8
dnspython==2.6.1
filelock==3.14.0
flasgger==0.9.7.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
idna==2.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jwt==1.3.1
MarkupSafe==2.1.5
mistune==3.0.2
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pip==24.0
platformdirs==4.2.2
pycparser==2.22
PyJWT==2.8.0
pymongo==4.7.2
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
six==1.16.0
urllib3==2.2.1
validators==0.28.1
virtualenv==20.26.2
Werkzeug==3.0.3